BOAC-2104, move dupe logic/styles in new-note and edit-note to manage-attachments
